more work on blogger

+++ Project2/Sources/circbuf.h
@@ -11,7 +11,7 @@
 #include <stdint.h>
 #include <stdlib.h>
 
-typedef uint32_t CircBufData_t;
+typedef uint8_t CircBufData_t;
 
 typedef struct CircBuf_t{
 	void* buffer;
